pyproject.toml
[tool.poetry]
name = "python-project-52"
version = "0.1.0"
description = "task manager"
authors = ["nickolay_kuzmin <nickolay.kuzmin2001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
python-dotenv = "^1.0.0"
dj-database-url = "^2.0.0"
django-bootstrap5 = "^23.3"
gunicorn = "^20.1.0"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
